refactor lambda directory structure
